requirements.txt
Django==2.1.2
django-extensions==2.1.3
django-filter==2.0.0
djangorestframework==3.9.0
Markdown==3.0.1
pygraphviz==1.5
pytz==2018.5
six==1.11.0
uWSGI==2.0.17.1
